requirements.txt
streamlit
torch
torchvision
timm
Pillow
opencv-python-headless
